ref/EV-760/fazendo a refatoração dos widgets
