+++ lib/app_widget.dart
@@ -1,4 +1,4 @@
-import 'package:crypto_app/src/presenter/pages/home_page.dart';
+import 'package:crypto_app/src/presenter/pages/portfolio/portfolio_page.dart';
 import 'package:device_preview/device_preview.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
