+++ lib/src/presenter/controllers/visibility/visibility_notifier.dart
@@ -3,7 +3,5 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 class VisibilityNotifier extends StateNotifier<bool> {
   VisibilityNotifier() : super(true);
 
-  void changeVisibility() {
-    state = !state;
-  }
+  void changeVisibility() => state = !state;
 }
