+++ lib/src/presenter/controllers/visibility/visibility_provider.dart
@@ -1,4 +1,5 @@
 import 'package:crypto_app/src/presenter/controllers/visibility/visibility_notifier.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-final visibilityProvider = StateProvider((ref) => VisibilityNotifier());
+final visibilityProvider = StateNotifierProvider<VisibilityNotifier, bool>(
+    (ref) => VisibilityNotifier());
