+++ lib/src/presenter/pages/portfolio/widgets/crypto_infos_in_tile.dart
@@ -1,0 +1,47 @@
+import 'package:flutter/material.dart';
+
+import '../../../../domain/entities/wallet_crypto_entity.dart';
+
+class CryptoInfosInTile extends StatelessWidget {
+  const CryptoInfosInTile({
+    Key? key,
+    required this.walletCryptoEntity,
+  }) : super(key: key);
+
+  final WalletCryptoEntity walletCryptoEntity;
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      children: [
+        CircleAvatar(
+          radius: 24,
+          backgroundImage: Image.asset(walletCryptoEntity.crypto.image).image,
+        ),
+        Padding(
+          padding: const EdgeInsets.only(left: 8),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Text(
+                walletCryptoEntity.crypto.initials,
+                style: const TextStyle(
+                  fontSize: 19,
+                  fontWeight: FontWeight.w500,
+                ),
+              ),
+              Text(
+                walletCryptoEntity.crypto.name,
+                style: TextStyle(
+                  fontSize: 15,
+                  color: Colors.grey.shade800,
+                  fontWeight: FontWeight.w400,
+                ),
+              ),
+            ],
+          ),
+        ),
+      ],
+    );
+  }
+}
